Add a select_boolean preset for fake booleans
This is a HACK, because ckan schemas returns boolean values as
strings, but the form select macro (rightly) expects booleans
to be booleans. I've added a new preset which attempts to deal with a
string "true|false" and casts it to a boolean in order to correctly set
and persist the value.

TODO: we need to find out why ckannext-scheming stores booleans as
strings and fix it.